로직 잔버그 수정

+++ palirates_server/palirates_server/GameObject.h
@@ -20,7 +20,7 @@ class GameObject : public std::enable_shared_from_this<GameObject>
 protected:
     Object_Type obj_type;
     string Obj_Name;
-    float Obj_ID;
+    int Obj_ID;
 
     shared_ptr<GameObject> child_obj = NULL;
     shared_ptr<GameObject> sibling_obj = NULL;
@@ -93,8 +93,8 @@ public:
     void Obj_Info(int depth = 0);
     Object_Type GetType() { return obj_type; }
 
-    void SetID(float ID) { Obj_ID = ID; }
-    float GetID() { return Obj_ID; }
+    void SetID(int ID) { Obj_ID = ID; }
+    int GetID() { return Obj_ID; }
 
     std::unordered_set<int> RootMotionTrackSet;
